# open-api-spec.yaml
openapi: 3.0.3
info:
  title: HSSE Tech API
  description: Описание REST API сервиса "HSSE Tech"
  version: 0.9.0
servers:
  - url: 'https'
paths:
...
  /api/renting/item:
    post:
      summary: "POST api/renting/item"
      operationId: "createItem"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateItemRequest"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Item"
    get:
      summary: "GET api/renting/item"
      operationId: "getAllItems"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/GetItemResponse"
  /api/renting/item/{item_id}/photo:
    post:
      summary: "POST api/renting/item/{item_id}/photo"
      operationId: "pinItemThumbnailPhoto"
      parameters:
        - name: "item_id"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
      responses:
        "200":
          description: "OK"
    get:
      summary: "GET api/renting/item/{item_id}/photo"
      operationId: "getItemThumbnailPhoto"
      parameters:
        - name: "item_id"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Resource"
  /api/renting/item/{id}:
    patch:
      summary: "PATCH api/renting/item/{id}"
      operationId: "updateItem"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateItemRequest"
        required: true
      responses:
        "204":
          description: "No Content"
  /api/renting/item/{itemId}:
    get:
      summary: "GET api/renting/item/{itemId}"
      operationId: "getItem"
      parameters:
        - name: "itemId"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
        - name: "loadRentInfo"
          in: "query"
          required: false
          schema:
            type: "boolean"
            default: "false"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/GetItemResponse"
    delete:
      summary: "DELETE api/renting/item/{itemId}"
      operationId: "deleteItem"
      parameters:
        - name: "itemId"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
      responses:
        "200":
          description: "OK"
  /api/renting/item/{item_id}/qr:
    get:
      summary: "GET api/renting/item/{item_id}/qr"
      operationId: "getItemBookingQRCode"
      parameters:
        - name: "item_id"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
      responses:
        "200":
          description: "OK"
          content:
            application/octet-stream:
              schema:
                $ref: "#/components/schemas/Resource"
  /api/renting/item/{item_id}/try-open:
    post:
      summary: "POST api/renting/item/{item_id}/try-open"
      operationId: "provideAccessToItemIfAllowed"
      parameters:
        - name: "item_id"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
      responses:
        "200":
          description: "OK"
...
components:
  schemas:
...
    BigDecimal:
      type: "object"
      properties: { }
    Instant:
      type: "object"
      properties: { }
...
    CreateItemRequest:
      type: "object"
      properties:
        display_name:
          type: "string"
        item_type_id:
          type: "string"
          format: "uuid"
    ItemType:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        cost:
          $ref: "#/components/schemas/BigDecimal"
        display_name:
          type: "string"
        max_rent_time_minutes:
          type: "integer"
          format: "int32"
    LockPassport:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        lockedItems:
          type: "array"
          items:
            $ref: "#/components/schemas/Item"
    Item:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        displayName:
          type: "string"
        type:
          $ref: "#/components/schemas/ItemType"
        lock:
          $ref: "#/components/schemas/LockPassport"
    Resource:
      type: "object"
      properties: { }
    UpdateItemRequest:
      type: "object"
      properties:
        new_display_name:
          type: "string"
    Rent:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        plannedStart:
          $ref: "#/components/schemas/Instant"
        plannedEnd:
          $ref: "#/components/schemas/Instant"
        item:
          $ref: "#/components/schemas/Item"
        factEnd:
          $ref: "#/components/schemas/Instant"
        factStart:
          $ref: "#/components/schemas/Instant"
        renter:
          $ref: "#/components/schemas/HumanUserPassport"
    Transaction:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        amount:
          $ref: "#/components/schemas/BigDecimal"
        name:
          type: "string"
        description:
          type: "string"
        createdAt:
          $ref: "#/components/schemas/Instant"
        wallet:
          $ref: "#/components/schemas/Wallet"
        status:
          type: "string"
          enum:
            - "IN_PROCESS"
            - "SUCCESS"
            - "FAILED"
    Wallet:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        owner:
          $ref: "#/components/schemas/HumanUserPassport"
        balance:
          $ref: "#/components/schemas/BigDecimal"
        transactions:
          type: "array"
          items:
            $ref: "#/components/schemas/Transaction"
    Role:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int32"
        name:
          type: "string"
        users:
          type: "array"
          items:
            $ref: "#/components/schemas/HumanUserPassport"
    HumanUserPassport:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        yandexId:
          type: "integer"
          format: "int64"
        firstName:
          type: "string"
        lastName:
          type: "string"
        isBanned:
          type: "boolean"
        email:
          type: "string"
        rents:
          type: "array"
          items:
            $ref: "#/components/schemas/Rent"
        wallet:
          $ref: "#/components/schemas/Wallet"
        roles:
          type: "array"
          items:
            $ref: "#/components/schemas/Role"
    GetShortRentResponse:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        renter:
          $ref: "#/components/schemas/HumanUserPassport"
        start_time:
          $ref: "#/components/schemas/Instant"
        end_time:
          $ref: "#/components/schemas/Instant"
    GetItemResponse:
      type: "object"
      properties:
        item_id:
          type: "string"
          format: "uuid"
        type_id:
          type: "string"
          format: "uuid"
        display_name:
          type: "string"
        rent_info:
          type: "array"
          items:
            $ref: "#/components/schemas/GetShortRentResponse"
...